+++ bindings/react/src/components/SimpleWrapper.jsx
@@ -4,12 +4,7 @@ import Util from './Util.js';
 
 class SimpleWrapper extends BasicComponent {
   render() {
-    var {...others} = this.props;
-
-    Util.convert(others, 'disabled');
-    Util.convert(others, 'ripple');
-
-    return React.createElement(this._getDomNodeName(), others, this.props.children);
+    return React.createElement(this._getDomNodeName(), Util.getAttrs(this), this.props.children);
   }
 }