+++ core/src/elements/ons-col.spec.js
@@ -9,18 +9,16 @@ describe('OnsColElement', () => {
     it('provides \'width\' attribute', () => {
       var element = ons._util.createElement('<ons-col width="100px"></ons-col>');
 
-      expect(element.style.webkitBoxFlex).to.equal('0');
-      expect(element.style.webkitFlex).to.equal('0 0 100px');
-      if ('mozBoxFlex' in element.style) {
-        expect(element.style.mozBoxFlex).to.equal('0');
-      }
-      if ('mozFlex' in element.style) {
+      if ('flex' in element.style) {
+        expect(element.style.flex).to.equal('0 0 100px');
+      } else if ('webkitFlex' in element.style) {
+        expect(element.style.webkitFlex).to.equal('0 0 100px');
+      } else if ('mozFlex' in element.style) {
         expect(element.style.mozFlex).to.equal('0 0 100px');
-      }
-      if ('msFlex' in element.style) {
+      } else if ('msFlex' in element.style) {
         expect(element.style.msFlex).to.equal('0 0 100px');
       }
-      expect(element.style.flex).to.equal('0 0 100px');
+
       expect(element.style.maxWidth ).to.equal('100px');
     });
   });
@@ -30,18 +28,16 @@ describe('OnsColElement', () => {
       var element = new ons.ColElement();
       element.setAttribute('width', '100px');
 
-      expect(element.style.webkitBoxFlex).to.equal('0');
-      expect(element.style.webkitFlex).to.equal('0 0 100px');
-      if ('mozBoxFlex' in element.style) {
-        expect(element.style.mozBoxFlex).to.equal('0');
-      }
-      if ('mozFlex' in element.style) {
+      if ('flex' in element.style) {
+        expect(element.style.flex).to.equal('0 0 100px');
+      } else if ('webkitFlex' in element.style) {
+        expect(element.style.webkitFlex).to.equal('0 0 100px');
+      } else if ('mozFlex' in element.style) {
         expect(element.style.mozFlex).to.equal('0 0 100px');
-      }
-      if ('msFlex' in element.style) {
+      } else if ('msFlex' in element.style) {
         expect(element.style.msFlex).to.equal('0 0 100px');
       }
-      expect(element.style.flex).to.equal('0 0 100px');
+
       expect(element.style.maxWidth ).to.equal('100px');
     });
   });