+++ core/src/elements/ons-page.js
@@ -32,8 +32,6 @@ const scheme = {
   '.page__background': 'page--*__background'
 };
 
-const nullToolbarElement = document.createElement('ons-toolbar'); // requires that 'ons-toolbar' element is registered
-
 /**
  * @element ons-page
  * @category page
@@ -227,26 +225,28 @@ export default class PageElement extends BaseElement {
   }
 
   connectedCallback() {
-    if (this._initialized || !util.isAttached(this)) { // Avoid detached calls
+    if (!util.isAttached(this)) { // Avoid detached calls
       return;
     }
 
-    this._initialized = true;
-
     contentReady(this, () => {
-      setImmediate(() => {
-        this.onInit && this.onInit();
-        util.triggerElementEvent(this, 'init');
-      });
-
-      if (!util.hasAnyComponentAsParent(this)) {
-        setImmediate(() => this._show());
-      }
-
       this._tryToFillStatusBar(); // Ensure status bar when the element was compiled before connected
 
       if (this.hasAttribute('on-infinite-scroll')) {
         this.attributeChangedCallback('on-infinite-scroll', null, this.getAttribute('on-infinite-scroll'));
+      }
+
+      if (!this._initialized) {
+        this._initialized = true;
+
+        setImmediate(() => {
+          this.onInit && this.onInit();
+          util.triggerElementEvent(this, 'init');
+        });
+
+        if (!util.hasAnyComponentAsParent(this)) {
+          setImmediate(() => this._show());
+        }
       }
     });
   }
@@ -354,7 +354,7 @@ export default class PageElement extends BaseElement {
   }
 
   _getToolbarElement() {
-    return util.findChild(this, 'ons-toolbar') || nullToolbarElement;
+    return util.findChild(this, 'ons-toolbar') || document.createElement('ons-toolbar');
   }
 
   static get observedAttributes() {